# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Events Search API",
        "description": "Events searchAPI documentation",
        "contact": {
            "email": "fgalvao90@yahoo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888/api",
            "description": "Events API Server"
        }
    ],
    "paths": {
        "/events/search": {
            "get": {
                "tags": [
                    "Authorize.Events"
                ],
                "summary": "Search the events",
                "description": "Returns list of events",
                "operationId": "ef4365e6361d7900f47ad4cb9799ca15",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query"
                    },
                    {
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "UnAuthorize",
            "description": "No user login required"
        },
        {
            "name": "Authorize",
            "description": "User login required"
        }
    ],
    "security": [
        []
    ]
}